modify: review 1~3 & seperate submitUserData Function
